PPTP-1191 Moved the partnership controllers to partner package
