+++ conf/app.routes
@@ -57,7 +57,7 @@ GET        /edit-address    uk.gov.hmrc.plasticpackagingtax.registration.control
 
 GET       /organisation-check-answers  uk.gov.hmrc.plasticpackagingtax.registration.controllers.organisation.CheckAnswersController.displayPage()
 
-GET        /partner-grs-callback   uk.gov.hmrc.plasticpackagingtax.registration.controllers.organisation.PartnerGrsController.grsCallback(journeyId)
+GET        /partner-grs-callback   uk.gov.hmrc.plasticpackagingtax.registration.controllers.partner.PartnerGrsController.grsCallback(journeyId)
 
 # Primary Contact Details
 GET        /main-contact-name    uk.gov.hmrc.plasticpackagingtax.registration.controllers.contact.ContactDetailsFullNameController.displayPage()
